Move a format_sources comment from unsupported position
kdev_format_source does not look for comments in the middle of a line.
It only discards lines that start with '#' (possibly after whitespaces).
This mistake did not result in a bug under GNU/Linux because the comment
was passed as a command to Bash, which ignored it:
    bash -c "# (otherwise this breaks unit tests)"
Let us fix this mistake so as not to mislead users that take KDevelop's
own format_sources as a syntax example.