kdev-clang: suspend translations units when we don't need them
This should significantly reduce the memory consumption of KDevelop
when lots of files are opened. Then, we used to keep the TUs for
these documents alive. Suspending them after we are done interpreting
the data basically gives us back some resources for free. We will
reparse the TU anyways before using it again the next time.

See upstream commit for some more context:
https://github.com/llvm-mirror/clang/commit/91321d9884b7ce357cac1d128244d22dbe2a0cc7