Actually remove temporary item repository directories
The shutdown procedure was changed in commit c3c9ea8a3c, but that
broke removal of temporary item repository directories. The repo
was shutdown before the session is cleared, which will try to
remove the dir. But it just marked the dir for removal, which should
then happen in shutdown, but that was called already. So detect that
and remove the dir directly then.