execute: remember the option to kill a process that is already running
We have the feature to kill the currently executed process as we run it.
This patch allows to remember the selection as I've seen myself doing
the same over and over again for different projects.

It leaves the same default behaviour and also adds an option in the
configuration dialog in case we need to change it back for some reason.