Clarify when GrepFindFilesThread::files() may be called
Calling this function concurrently with GrepFindFilesThread::run() would
be a data race. Fortunately, it is called only in
GrepJob::slotFindFinished(), which is connected to
&GrepFindFilesThread::finished.

Although I could not find this spelled out in Qt documentation, I think
it is safe to access shared data without locking in a slot connected to
QThread::finished(). It looks like the process of sending a signal from
one thread and delivering it to a slot in another thread employs
acquire and release semantics. The QThread::finished() documentation
specifies that "This signal is emitted from the associated thread right
before it finishes executing" unless "the associated thread was
terminated using terminate()". My examination of relevant Qt 5.15 code
revealed that:
  1) QThread::finished() signal is emitted after QThread::run() returns;
  2) before passing the signal to the receiver's thread, the sender's
     thread locks receiver->d_func()->threadData->postEventList.mutex in
     QCoreApplication::postEvent();
  3) the receiver's thread locks the same mutex in
     QCoreApplicationPrivate::sendPostedEvents() before calling the
     connected slot.

Eliminate an unnecessary copying in files() to improve performance and
rename the member function to takeFiles() to match its new semantics.