Remove "text/x-diff" alias from supported MIME types
"text/x-diff" is an alias for "text/x-patch" MIME type. Therefore its
glob patterns are identical: "*.diff" and "*.patch". The downside of the
presence of this alias in the SupportedMimeTypes list is that
LanguageController must either match all filenames against duplicate
patterns (as it does now) or detect and remove these duplicate patterns,
which would entail unnecessary extra code and still some run-time work.

There are many "x-patch" uses in KDevelop code base, but this was the
only use of "x-diff". That is why I don't think this alias will be
missed by anyone.