Check ICore::activeSession() for nullptr in BreakpointModel::save()
When a user exits KDevelop while debugging a program, a queued call to
BreakpointModel::save() may be invoked during Core::cleanup() or perhaps
even ~CorePrivate(). If this happens between
SessionController::cleanup() and ~DebugController() (which is the parent
of BreakpointModel), the result is a crash. A breakpoint change that is
saved at this time is likely unimportant and can be safely skipped.

BUG: 425986
FIXED-IN: 5.6.1