Do not open documents after DocumentController::cleanup()
For now d->shuttingDown is checked in just one of many
DocumentController::openDocument() overloads and alternatives, because
this is the only overload that is known to be called after cleanup() and
cause crashes. If similar crashes happen because of other ways of
opening a document, the check can be moved or reused elsewhere.
Or perhaps a different strategy will be needed to fix these hypothetical
future crashes: e.g. simply returning nullptr doesn't work if the caller
dereferences the returned document pointer without checking.

BUG: 425994
FIXED-IN: 5.6.1