Don't show a busy indicator for 5s after a task completion
81220db11cb17e0ffa4499bec34dfb9e21ca1aa2 eliminated such misleading
progress reports for items that use a busy indicator themselves, but
missed the case when a quick percent-progress-reporting task follows
right after a busy-indicator task.

For example, GrepJob's WorkCollectFiles step uses a busy indicator, then
its next WorkGrep step reports progress in percents. So when the
WorkGrep step is completed in less than a second (e.g. when there are
not many files to search in and the looked-for string is not present in
any of them), the bottom-right-corner progress bar shows
a busy indicator for 5 seconds.

Let us display 100% progress during these 5 seconds instead.