Default timeout in showErrorMessage(): [01]s => 5s
The default timeout value was 0 seconds in showErrorMessage() signal of
IStatus and derived classes that override it. The default timeout value
was 1 second in IUiController::showErrorMessage() member function.

In all the three places, where classes derived from IStatus emit this
signal, no timeout is specified, meaning that the error messages were
shown, then hidden in 0 seconds, and so were never actually visible to
the left of the progress bar in the bottom-right corner. The calls to
IUiController's member function never rely on the default timeout value.
However, since both emitting of this signal and calls to this function
are eventually handled by StatusBar::showErrorMessage(), it makes sense
to use the same default timeout value. The user will have a chance to
actually see and read the error message if it is visible for 5 seconds,
while 1 second is probably not long enough for this.

Remove default timeout values from overriden signals that are never
emitted. Some of the derived classes already omit this default value in
overriden signals that are never emitted. Let us increase the number of
such cases. This will make it easier to adjust the default timeout value
in the future. If someone decides to emit one of these signals at some
future time, they can copy the default value from IStatus.

Remove unused timeout parameter from
GrepOutputModel::showErrorMessage(). Only one slot
GrepOutputView::showErrorMessage(const QString& errorMessage) is
connected to this signal.