Make CMake::FileApi::ImportJob Killable
This commit fixes the following bug: when a user opens a CMake project
and then immediately hits the Stop All toolbar button while the project
is being loaded, the loading process is stuck until KDevelop exit. The
visible effects of this failed job killing are:
1) a busy indicator in the bottom-right status progress bar with the
   text "Loading Project <project name>";
2) an unstoppable job in the {main menu => Run => Stop submenu} called
   "<KJob> Unnamed job";
3) an attempt to open the stuck-at-loading project again fails with a
   warning message at the top of the KDevelop main window: "Already
   opening <path to the project's kdev4 file>, not opening again".

An unfortunate invisible effect of this failed job killing is that
neither &IProjectController::projectOpeningAborted nor
&IProjectController::projectOpened signal is emitted after
&IProjectController::projectAboutToBeOpened, which causes several leaks
in the objects that listen to these signals.

Making CMake::FileApi::ImportJob Killable allows to kill the
ExecuteCompositeJob that contains it as a subjob, and lets the
&IProjectController::projectOpeningAborted signal be emitted.

The QFuture returned by QtConcurrent::run() cannot be canceled after the
task is started. But calling `m_futureWatcher.cancel()` is still
beneficial in two ways:
1) prevents starting a not-yet-started task;
2) allows to skip emitting signals from a lambda connected to
   &QFutureWatcher<CMakeProjectData>::finished if
   m_futureWatcher.isCanceled() returns true.

I think it is safe to let the lambda passed to QtConcurrent::run() in
CMake::FileApi::ImportJob::start() run to completion after the job is
killed. The lambda doesn't access the job object and does not seem to
reference the KDevelop core, so there is no need for any extra checks.