Don't call `abortOpeningProject()` when project reloading fails
Currently ProjectController::abortOpeningProject() has only two effects:
1) removes the project file URL from
   ProjectControllerPrivate::m_currentlyOpening list;
2) emits &IProjectController::projectOpeningAborted, to which only one
   slot BackgroundParser::projectOpeningAborted() is connected, and
   which removes the project from
   BackgroundParserPrivate::m_loadingProjects set.

A project is not inserted into either of the two containers when it is
reloaded, so both effects didn't actually change anything.

However, more slots should be connected to the
&IProjectController::projectOpeningAborted signal. And once they get
connected, they wouldn't want this signal to be emitted when reloading a
project fails.