Project file managers: listen to projectOpeningAborted()
When opening a project is aborted, &IProjectController::projectClosing
signal is never emitted and a pointer to the destroyed project
along with other associated objects remains in project file managers'
data structures until the application exit.

AbstractFileManagerPlugin::import(), CustomMakeManager::import() are
called and CMakeManager::integrateData() is normally[1] invoked before
projectController emits either projectOpened() (eventually followed by
projectClosing()) or projectOpeningAborted() signal. All references to
an about-to-be-destroyed project object must be removed in one of the
three projectClosing() slots now connected to both projectClosing() and
projectOpeningAborted() signals. Such references may not exist when
projectController emits the projectOpeningAborted() signal (in which
case the aforementioned import()/integrateData() are never invoked), so
the projectClosing() slot implementations must (and currently do) handle
this gracefully.

[1] If CMake::FileApi is used successfully, then
CMakeManager::integrateData() can not possibly be invoked after
projectController emits projectOpeningAborted() or projectOpened(),
because CMake::FileApi::ImportJob emits dataAvailable() (which results
in a synchronous call to integrateData()) strictly before finished() and
result() (which happen before projectController possibly emits
projectOpeningAborted()). However, if tryCMakeServer() is called, then
there is a possibility of the wrong order of invocations => a leak or a
crash. Since CMake server mode is deprecated and normally not used if
the CMake version is >= 3.14, I don't want to experiment and test it,
especially because the changes in this commit won't make the possible
bugs any worse.

Replace ICore::self()->projectController() with
core()->projectController() as the latter seems to be the proper way to
obtain a pointer to Core from plugins.