AbstractFileManagerPlugin: don't leak KDirWatch objects
When Project::reloadModel() is called because of a CMake configuration
change, AbstractFileManagerPlugin::import() is called with a project
that is already present in AbstractFileManagerPlugin's m_watchers and
m_filters. The code doesn't detect this, creates a new KDirWatch and
replaces the existing pointer in m_watchers. The old watcher is leaked
and is not destroyed until its parent - the project - is destroyed.
The same filters are created and stored in m_filters, but this is not a
leak because the filters' value type is QVector<Filter>, not a pointer.

This leak does not just waste CPU time creating new objects, but also
keeps the duplicate KDirWatch object alive and monitoring the project
directory. Furthermore, lambdas defined in CMakeManager::integrateData()
remain connected to this old leaked KDirWatch, not its replacement. The
leaked KDirWatch is destroyed not when
&IProjectController::projectClosing signal is emitted, but later - when
the project object is destroyed. This can trigger assertions in the
aforementioned lambdas, or cause a crash if assertions are disabled,
because Project::projectItem() is destroyed when a project is closed.

Let us detect when the project is already present in m_watchers and
m_filters. Return early instead of creating identical new objects and
replacing the good old ones.