TestView: listen to projectOpeningAborted()
When opening a project is aborted after some of its test suites were
added, &IProjectController::projectClosed signal is never emitted and
the project item with its test suites remain in Unit Tests tool view.
When the same project is opened again, duplicate test suites are visible
in the Unit Tests tool view.

Note that this fix does not guarantee that no test suites for the not
opened project remain: if the &IProjectController::projectOpeningAborted
signal is emitted after test suites are created but before all of them
are added, ...

TODO: deal with this non-guarantee before this commit!!!