Add exception specification to function signatures
When using the IDE to generate function definitions, some elements
from the function signature are omitted. One of those are exception
specifications. This commit adds exception specification to function
signatures if the specification is `noexcept` (no expression
 provided) or `throw()` (no types provided). More complex
specifications cannot be added for now as libclang does not provide
the required information.

CCBUG: 188180