Don't delay Stop and Tool Views action menus
For many years I thought that KDevelop's Stop toolbar menu was broken.
Today after digging into the relevant code and documentation I found out
that one has to press and hold the button for a few seconds for the
submenu to show up.

From the documentation to KActionMenu::setDelayed():
  Remember that if the "main" action (the toolbar button itself) cannot
  be clicked, then you should call setDelayed(false).

This is clearly the case for the Stop menu button and for the Tool Views
menu button (which can be added to the Main Toolbar through Configure
Toolbars dialog): clicking these buttons has no effect.