Make the global variable doReferenceCounting thread_local
This variable is accessed concurrently from multiple threads. It is
written to under a mutex lock, but read from without any thread
synchronization. Therefore the variable has to be either atomic or
thread_local to prevent data races.

As far as I can tell, the variable's value is never intended to be
shared with other threads. If one thread temporarily sets
doReferenceCounting to true, and then another thread reads the variable
before the first thread resets it to false, there is redundant mutex
locking. If this other thread then (redundantly) sets
doReferenceCounting to true, performs some operation and resets it to
false; then the first thread reads this variable again, some necessary
checks and steps can be skipped in the first thread. So thread_local
seems to be the more correct and performant alternative.

This change does not break any kdevelop or kdev-python tests.