Fix test_gdbprinters by making it more resilient
Don't rely on `next` bringing us to the expected line. Instead,
request the line explicitly and continue until we land there.
Apparrently some compiler change triggered a failure of the
test then otherwise, as we ended up in the A::A ctor in line 11
while we actually wanted to print the pointerList in loc 46.