Reduce time required to run bench_quickopen
By doing more :) QTestLib otherwise detecs that the time required
for a single loop is too fast, it then retries with a higher iteration
count but that requires running the setup code again. And that is
actually way slower than the stuff we want to benchmark here...

So increase the workload of the benchmark code, or altneratively use
QBENCHMARK_ONCE with a large inner loop count.

Previously, the whole benchmark took ~25s on my system. Now it only
takes ~7s and gives better data too.