Don't leak deleted breakpoints
But really, this is just a stop-gap measure to
prevent LSAN reports in the unit test. We still
basically leak the deleted breakpoints until we
destroy the model, so nothing really is won here.
But it seems like the whole ownership and lifetime
tracking of breakpoints is utterly broken. Note how
we store e.g. raw pointers in the breakpoint controller
and don't seem to clean those up ever properly either...